requirements.txt
Flask==2.2.5


matplotlib==3.7.1

numpy==1.26.4


pandas==2.2.2


pillow==10.4.0

scikit-learn==1.5.2
scipy==1.13.1

seaborn==0.13.1


tensorflow==2.17.0
